requirements.txt
Appium-Python-Client==5.2.0
selenium==4.34.2
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
webdriver-manager==4.0.1
python-dotenv==1.0.0
allure-pytest==2.13.2
pytest-json-report==1.5.0
